backend/requirements-backend.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
opencv-python-headless
tensorflow
keras
numpy<2.0
pydantic
